python/agents/knowledge_retrieval_challenge/requirements.txt
google-genai
faiss-cpu
pdf2image
Pillow
pydantic
python-dotenv
rank_bm25
tenacity
